Fix demo
