+++ demo/package.json
@@ -10,7 +10,7 @@
   "devDependencies": {
     "@babel/plugin-proposal-class-properties": "^7.0.0-beta.55",
     "@babel/plugin-proposal-decorators": "^7.4.0",
-    "vite": "^4.2.1"
+    "vite": "^5.4.10"
   },
   "dependencies": {
     "@material-ui/core": "4.9.5",
